+++ src/main/java/com/prgrms/rg/domain/ridingpost/model/RidingPostFinder.java
@@ -3,6 +3,8 @@ package com.prgrms.rg.domain.ridingpost.model;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.prgrms.rg.domain.ridingpost.model.exception.RidingPostNotFoundException;
+
 import lombok.RequiredArgsConstructor;
 
 @Component
