+++ src/main/java/com/prgrms/rg/domain/ridingpost/model/exception/RidingPostNotFoundException.java
@@ -1,4 +1,4 @@
-package com.prgrms.rg.domain.ridingpost.model;
+package com.prgrms.rg.domain.ridingpost.model.exception;
 
 public class RidingPostNotFoundException extends RuntimeException {
 	public RidingPostNotFoundException(Long id) {
